Update slotmachine.ino
This is the code for a slot machine which can be displayed in the Adafruit ssd1306 32x128 i2c oled display, in conjunction with a simple joystick, pushbutton, Arduino microcontroller and an optional RGB led for lighting effects.  I wrote is all myself despite having little to no programming background.  It may be longer than it needs to be.  It uses 99% of an Arduino Nano's storage capacity.